+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Changelog
 
+- Node.js 4 support is dropped (breaking change).
 - New transformation functions are added: `radiansToDegrees`, `degreesToRadians`, `radiansLat`, `radiansLong`.
 
 ### 2.0.3 (2018-09-15)